Revert "changed type of PKI status from 'long' to 'int', e.g., in CMP_CTX_status_get()"
This reverts commit 5954c5f26881ca94d2d32928ff47133329fc9d45.

+++ crypto/cmp/cmp_int.h
@@ -84,7 +84,7 @@ struct cmp_ctx_st {
     int ignore_keyusage; /* ignore key usage entry in certs */
     int maxPollTime; /* maximum number of seconds to poll the server
                         for a response if a 'waiting' PKIStatus is received */
-    int lastPKIStatus; /* PKIStatus of last received IP/CP/KUP/RP, or -1 */
+    long lastPKIStatus; /* PKIStatus of last received IP/CP/KUP/RP, or -1 */
     /* TODO: this should be a stack since there could be more than one */
     unsigned long failInfoCode; /* failInfoCode of last received IP/CP/KUP */
     /* TODO: this should be a stack since there could be more than one */
