+++ include/openssl/cmp.h
@@ -379,7 +379,7 @@ CMP_INFOTYPEANDVALUE *CMP_ITAV_new(const ASN1_OBJECT *type,
                                    const ASN1_TYPE *value);
 CMP_PKISTATUSINFO *CMP_statusInfo_new(int status, int failure,
                                       const char *text);
-int CMP_PKISTATUSINFO_PKIStatus_get(CMP_PKISTATUSINFO *statusInfo);
+long CMP_PKISTATUSINFO_PKIStatus_get(CMP_PKISTATUSINFO *statusInfo);
 
 X509 *CMP_CERTRESPONSE_get_certificate(CMP_CTX *ctx, CMP_CERTRESPONSE *crep);
 int CMP_PKIFAILUREINFO_check(ASN1_BIT_STRING *failInfo, int codeBit);
@@ -504,7 +504,7 @@ int CMP_CTX_set0_reqExtensions(CMP_CTX *ctx, X509_EXTENSIONS *exts);
 int CMP_CTX_set1_serverPath(CMP_CTX *ctx, const char *path);
 int CMP_CTX_set_failInfoCode(CMP_CTX *ctx, CMP_PKIFAILUREINFO *failInfo);
 unsigned long CMP_CTX_failInfoCode_get(CMP_CTX *ctx);
-int CMP_CTX_status_get(CMP_CTX *ctx);
+long CMP_CTX_status_get(CMP_CTX *ctx);
 CMP_PKIFREETEXT *CMP_CTX_statusString_get(CMP_CTX *ctx);
 # define CMP_CTX_OPT_MSGTIMEOUT                 0
 # define CMP_CTX_OPT_MAXPOLLTIME                1
